move ScriptSequence into its own module
