+++ cli/src/cmd/forge/script/receipts.rs
@@ -1,6 +1,7 @@
 use std::sync::Arc;
 
-use crate::{cmd::ScriptSequence, init_progress, update_progress, utils::print_receipt};
+use super::sequence::ScriptSequence;
+use crate::{init_progress, update_progress, utils::print_receipt};
 use ethers::prelude::{Http, PendingTransaction, Provider, RetryClient, TxHash};
 use futures::StreamExt;
 use indicatif::{ProgressBar, ProgressStyle};
